ai-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pytesseract==0.3.10
Pillow==10.0.0
transformers==4.35.0
torch==2.1.0
keybert==0.8.3
prophet==1.1.4
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0
